Update Viewer.html
I wanted to test this to since this script is used by me for android. I didn't know if anyone is using it but me and if not i wanted to test this out just to see what this would do. The css fixes for the viewer.css seems to be hard to fix due to the multiple devices that use it. I figured since I'm using this then i can test it out with my stuff to see if this would fix android issues.