Update Viewer.html
Made a mistake on the the first commit. Here is the fix for android if no one is using the script that we made for android tv. I wanted to test to see if this would fix the css issues with android since we can't get anything to work for android and others without having issues.