Update videoScripts.js
I would like to have the hidden function moved to the top in the pause process. The seek and pause causes android to show a grey screen with a play button at the end of every video thats being played. Moving the hidden function to the top would prevent that from being shown.