Update Viewer.html
Please set this new change on the test server. The last change I noticed that the default meta tag never got posted if nothing is in the url. I made the correct changes and should work now. If my test go ok is there a way this can be added to the main sometime this week. Thanks