Update Viewer.html
I did a few more test and decided on this script. The reason for the shorter version is that i figure it be best for me to pass what i need to the viewport from my app. That way my app controls the size at any given point. It also will keep me from adding things in the future as well to the viewer. This way if i need to change something i can do it from the app. :) Anyway this is what i need on the test server now. Thanks